+++ lib/sponsors.ts
@@ -1,5 +1,4 @@
-import { env } from '@/env';
-import { getGitHubSponsors } from '@/lib/github';
+import { env } from '../env';
 
 export const GITHUB_SPONSOR_USERNAME = 'wotschofsky';
 
@@ -8,6 +7,27 @@ export type Sponsor = {
   name: string;
   logoUrl: string;
   url: string;
+};
+
+type SponsorsQueryResponse = {
+  data: {
+    user: {
+      sponsorshipsAsMaintainer: {
+        nodes: {
+          sponsorEntity: {
+            login: string;
+            name: string;
+            url: string;
+            avatarUrl: string;
+            websiteUrl: string | null;
+          };
+          tier: {
+            id: string;
+          } | null;
+        }[];
+      };
+    };
+  };
 };
 
 const buildSponsorUrl = (baseUrl: string) => {
@@ -20,20 +40,100 @@ const buildSponsorUrl = (baseUrl: string) => {
   }
 };
 
-export const getAllSponsors = async (): Promise<Sponsor[]> => {
-  const githubSponsors = await getGitHubSponsors(GITHUB_SPONSOR_USERNAME);
+const fetchGitHubSponsors = async () => {
+  if (!env.GITHUB_TOKEN) {
+    return [];
+  }
 
-  const sponsors = [
+  const response = await fetch('https://api.github.com/graphql', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${env.GITHUB_TOKEN}`,
+    },
+    body: JSON.stringify({
+      query: `
+        query SponsorsQuery($username: String!) {
+          user(login: $username) {
+            sponsorshipsAsMaintainer(first: 100) {
+              nodes {
+                sponsorEntity {
+                  ... on User {
+                    login
+                    name
+                    url
+                    avatarUrl
+                    websiteUrl
+                  }
+                  ... on Organization {
+                    login
+                    name
+                    url
+                    avatarUrl
+                    websiteUrl
+                  }
+                }
+                tier {
+                  id
+                }
+              }
+            }
+          }
+        }
+      `,
+      variables: {
+        username: GITHUB_SPONSOR_USERNAME,
+      },
+    }),
+  });
+
+  if (!response.ok) {
+    const body = await response.text();
+    throw new Error(
+      `Failed to fetch GitHub sponsors: ${response.status} ${response.statusText}\n${body}`,
+    );
+  }
+
+  const body = (await response.json()) as SponsorsQueryResponse;
+
+  const sponsors = body.data.user.sponsorshipsAsMaintainer.nodes.map((node) => {
+    if (!node.tier) {
+      throw new Error(
+        'Failed to fetch GitHub sponsors; access token might lack sufficient permissions',
+      );
+    }
+
+    return {
+      login: node.sponsorEntity.login,
+      name: node.sponsorEntity.name,
+      url: node.sponsorEntity.url,
+      avatarUrl: node.sponsorEntity.avatarUrl,
+      websiteUrl: node.sponsorEntity.websiteUrl,
+      tierId: node.tier.id,
+    };
+  });
+
+  return env.GITHUB_SPONSORS_FEATURED_TIERS
+    ? sponsors.filter((sponsor) =>
+        env.GITHUB_SPONSORS_FEATURED_TIERS!.includes(sponsor.tierId),
+      )
+    : sponsors;
+};
+
+export const getAllSponsors = async (): Promise<Sponsor[]> => {
+  const githubSponsors = await fetchGitHubSponsors();
+
+  const allSponsors = [
     ...(env.SPONSORS ?? []),
-    ...githubSponsors.map((s) => ({
-      id: s.login,
-      name: s.name,
-      logoUrl: s.avatarUrl,
-      url: s.websiteUrl || s.url,
+    ...githubSponsors.map((sponsor) => ({
+      id: sponsor.login,
+      name: sponsor.name,
+      logoUrl: sponsor.avatarUrl,
+      url: sponsor.websiteUrl || sponsor.url,
     })),
   ];
 
-  const sponsorsWithRef = sponsors.map((sponsor) => ({
+  const sponsorsWithRef = allSponsors.map((sponsor) => ({
     ...sponsor,
     url: buildSponsorUrl(sponsor.url),
   }));
